Mon, Feb 27, 2023, 1:01 PM +01:00

+++ CS50/Week 2/readability/readability.c
@@ -33,7 +33,7 @@ int main(void)
     }
     else
     {
-        printf("Grade %i\n", round(index));
+        printf("Grade %i\n", (int) round(index));
     }
 }
